backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
supabase
openai
datasets
transformers
torch
scikit-learn
pandas
numpy
pytest
pytest-asyncio
pytest-cov
httpx
black
isort
flake8
mypy
